Building the relationship between images to items
Items know what images it has, but images do not know what is its host. With this, it will go through each bag and associate its images to the appropriate item. There may be a better way to do this association than the 'Try Except' method I used. I also made the assumption that the item name is the same as the name of the first image it contains without the ending letters.

While looking at the number of items vs images, I found that there are 57 more images than items, something which I feel is wrong. I think that if there is a substantial amount of images to items, then this number should be far greater.